rename processing algos
