+++ enmapbox/apps/SpecDeepMap/__init__.py
@@ -1,12 +1,12 @@
 from enmapbox.gui.applications import EnMAPBoxApplication
 
 try:
-    from enmapbox.apps.SpecDeepMap.processingalgorithm_dataset_maker import DatasetMaker
-    from enmapbox.apps.SpecDeepMap.processingalgorithm_raster_splitter import RasterSplitter
-    from enmapbox.apps.SpecDeepMap.processingalgorithm_deep_learning_trainer import DL_Trainer
-    from enmapbox.apps.SpecDeepMap.processingalgorithm_deep_learning_mapper import DL_Mapper
-    from enmapbox.apps.SpecDeepMap.processingalgorithm_tensorboard_visualizer import Tensorboard_visualizer
-    from enmapbox.apps.SpecDeepMap.processingalgorithm_tester import DL_Tester
+    from enmapbox.apps.SpecDeepMap.processing_algorithm_dataset_maker import DatasetMaker
+    from enmapbox.apps.SpecDeepMap.processing_algorithm_raster_splitter import RasterSplitter
+    from enmapbox.apps.SpecDeepMap.processing_algorithm_deep_learning_trainer import DL_Trainer
+    from enmapbox.apps.SpecDeepMap.processing_algorithm_deep_learning_mapper import DL_Mapper
+    from enmapbox.apps.SpecDeepMap.processing_algorithm_tensorboard_visualizer import Tensorboard_visualizer
+    from enmapbox.apps.SpecDeepMap.processing_algorithm_tester import DL_Tester
 
     wrongEnv = False
 except Exception as ex:
