+++ enmapbox/apps/SpecDeepMap/processing_algorithm_dataset_maker.py
